// react-forms/src/components/Form/Form.enums.ts
export enum ErrorKeys {
  name = 'name',
  surname = 'surname',
  date = 'date',
  country = 'country',
  consent = 'consent',
  image = 'image',
}

export enum ErrorMessages {
  lengthError = 'length must be greater than',
  wrongDate = 'too young',
  emptyDate = 'enter your date of birth',
  country = 'choose your country',
  consent = 'confirm an agreement',
  image = 'choose an image',
}

export enum Gendors {
  female = 'Female',
  male = 'Male',
}

export enum LabelTexts {
  name = 'Name',
  surname = 'Surname',
  date = 'Birthday',
  consent = 'I agree to my personal data being processed',
}

export enum InputTypes {
  text = 'text',
  date = 'date',
  checkbox = 'checkbox',
  submit = 'submit',
  file = 'file',
}

export enum InputNames {
  name = 'name',
  surname = 'surname',
  date = 'date',
  consent = 'consent',
  gender = 'gender',
  country = 'country',
  image = 'image',
}

export enum Countries {
  default = 'default',
  Belarus = 'Belarus',
  Ukraine = 'Ukraine',
  Poland = 'Poland',
  Russia = 'Russia',
  Lithuania = 'Lithuania',
  Latvia = 'Latvia',
}

export enum TestIds {
  form = 'form',
  formCard = 'formCard',
  formCardList = 'formCardList',
  name = 'name',
  surname = 'surname',
  date = 'date',
  image = 'image',
  error = 'error',
  consent = 'consent',
  country = 'country',
  submit = 'submit',
  successMessage = 'successMessage',
  gender = 'gender',
}